VarID2 Analysis
This repository contains R source code for reproducing the VarID2 analysis performed in Rosales-Alvarez et al., (2023)

VarID is part of the RaceID package starting with RaceID v0.2.4 The scripts in this repository are designed for RaceID v0.2.6

Several data sets were analysed. For each of them there is one individual .R script. Refer to the main figures of the manuscript for further details.